backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
google-generativeai==0.7.2
httpx==0.27.0
sqlalchemy==2.0.30
asyncpg==0.29.0
supabase==2.5.0
upstash-redis==1.1.0
python-dotenv==1.0.1
tiktoken==0.7.0
pydantic==2.7.1
pydantic-settings==2.3.1
sse-starlette==2.1.0
python-multipart==0.0.9
openai
